# openapi-8001.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/me":{"get":{"summary":"Get Me","operationId":"get_me_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}}}},"/users/":{"post":{"summary":"Create User","operationId":"create_user_users__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Read Users","operationId":"read_users_users__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRead"},"title":"Response Read Users Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"put":{"summary":"Update User","operationId":"update_user_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Read User","operationId":"read_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user_results/":{"post":{"summary":"Create User Result","operationId":"create_user_result_user_results__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResultCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResultRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/average_data/":{"post":{"summary":"Create Average Data","operationId":"create_average_data_average_data__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AverageDataCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AverageDataResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/average_data/grade/{grade}":{"get":{"summary":"Read Average Data By Grade","operationId":"read_average_data_by_grade_average_data_grade__grade__get","parameters":[{"name":"grade","in":"path","required":true,"schema":{"type":"string","title":"Grade"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AverageDataResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/max_data/":{"post":{"summary":"Create Max Data","operationId":"create_max_data_max_data__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaxDataCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaxDataResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/max_data/grade/{grade}":{"get":{"summary":"Read Max Data By Grade","operationId":"read_max_data_by_grade_max_data_grade__grade__get","parameters":[{"name":"grade","in":"path","required":true,"schema":{"type":"string","title":"Grade"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaxDataResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user_results/{user_id}":{"get":{"summary":"Read User Results","operationId":"read_user_results_user_results__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResultRead"},"title":"Response Read User Results User Results  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/flexibility-checks/":{"get":{"summary":"Read Flexibility Checks","operationId":"read_flexibility_checks_flexibility_checks__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FlexibilityCheckRead"},"type":"array","title":"Response Read Flexibility Checks Flexibility Checks  Get"}}}}}},"post":{"summary":"Create Flexibility Check","operationId":"create_flexibility_check_flexibility_checks__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlexibilityCheckCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlexibilityCheckRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/flexibility-checks/{check_id}":{"put":{"summary":"Update Flexibility Check","operationId":"update_flexibility_check_flexibility_checks__check_id__put","parameters":[{"name":"check_id","in":"path","required":true,"schema":{"type":"integer","title":"Check Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlexibilityCheckCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlexibilityCheckRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Flexibility Check","operationId":"delete_flexibility_check_flexibility_checks__check_id__delete","parameters":[{"name":"check_id","in":"path","required":true,"schema":{"type":"integer","title":"Check Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user_results/{result_id}":{"delete":{"summary":"Delete User Result","operationId":"delete_user_result_user_results__result_id__delete","parameters":[{"name":"result_id","in":"path","required":true,"schema":{"type":"integer","title":"Result Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/average_max_data/":{"post":{"summary":"Create Average Max Data","operationId":"create_average_max_data_average_max_data__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AverageMaxDataCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AverageMaxDataRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/average_max_data/grade/{grade}":{"get":{"summary":"Read Average Max Data By Grade","operationId":"read_average_max_data_by_grade_average_max_data_grade__grade__get","parameters":[{"name":"grade","in":"path","required":true,"schema":{"type":"string","title":"Grade"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AverageMaxDataRead"},"title":"Response Read Average Max Data By Grade Average Max Data Grade  Grade  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AverageDataCreate":{"properties":{"grade":{"type":"string","title":"Grade"},"long_jump_cm":{"type":"number","title":"Long Jump Cm"},"fifty_meter_run_ms":{"type":"number","title":"Fifty Meter Run Ms"},"spider_ms":{"type":"number","title":"Spider Ms"},"eight_shape_run_count":{"type":"number","title":"Eight Shape Run Count"},"ball_throw_cm":{"type":"number","title":"Ball Throw Cm"},"total_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Total Score"}},"type":"object","required":["grade","long_jump_cm","fifty_meter_run_ms","spider_ms","eight_shape_run_count","ball_throw_cm"],"title":"AverageDataCreate"},"AverageDataResponse":{"properties":{"grade":{"type":"string","title":"Grade"},"long_jump_cm":{"type":"number","title":"Long Jump Cm"},"fifty_meter_run_ms":{"type":"number","title":"Fifty Meter Run Ms"},"spider_ms":{"type":"number","title":"Spider Ms"},"eight_shape_run_count":{"type":"number","title":"Eight Shape Run Count"},"ball_throw_cm":{"type":"number","title":"Ball Throw Cm"},"total_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Total Score"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["grade","long_jump_cm","fifty_meter_run_ms","spider_ms","eight_shape_run_count","ball_throw_cm","id"],"title":"AverageDataResponse"},"AverageMaxDataCreate":{"properties":{"grade":{"type":"string","title":"Grade"},"long_jump":{"type":"number","title":"Long Jump"},"fifty_meter_run":{"type":"number","title":"Fifty Meter Run"},"spider":{"type":"number","title":"Spider"},"eight_shape_run":{"type":"integer","title":"Eight Shape Run"},"ball_throw":{"type":"number","title":"Ball Throw"},"total_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Score"}},"type":"object","required":["grade","long_jump","fifty_meter_run","spider","eight_shape_run","ball_throw"],"title":"AverageMaxDataCreate"},"AverageMaxDataRead":{"properties":{"grade":{"type":"string","title":"Grade"},"long_jump":{"type":"number","title":"Long Jump"},"fifty_meter_run":{"type":"number","title":"Fifty Meter Run"},"spider":{"type":"number","title":"Spider"},"eight_shape_run":{"type":"integer","title":"Eight Shape Run"},"ball_throw":{"type":"number","title":"Ball Throw"},"total_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Score"},"id":{"type":"integer","title":"Id"},"type":{"type":"string","title":"Type"}},"type":"object","required":["grade","long_jump","fifty_meter_run","spider","eight_shape_run","ball_throw","id","type"],"title":"AverageMaxDataRead"},"FlexibilityCheckCreate":{"properties":{"title":{"type":"string","title":"Title"},"image_path":{"type":"string","title":"Image Path"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","image_path","description"],"title":"FlexibilityCheckCreate"},"FlexibilityCheckRead":{"properties":{"title":{"type":"string","title":"Title"},"image_path":{"type":"string","title":"Image Path"},"description":{"type":"string","title":"Description"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["title","image_path","description","id","created_at","updated_at"],"title":"FlexibilityCheckRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MaxDataCreate":{"properties":{"grade":{"type":"string","title":"Grade"},"long_jump_cm":{"type":"integer","title":"Long Jump Cm"},"fifty_meter_run_ms":{"type":"integer","title":"Fifty Meter Run Ms"},"spider_ms":{"type":"integer","title":"Spider Ms"},"eight_shape_run_count":{"type":"integer","title":"Eight Shape Run Count"},"ball_throw_cm":{"type":"integer","title":"Ball Throw Cm"},"total_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Score"}},"type":"object","required":["grade","long_jump_cm","fifty_meter_run_ms","spider_ms","eight_shape_run_count","ball_throw_cm"],"title":"MaxDataCreate"},"MaxDataResponse":{"properties":{"grade":{"type":"string","title":"Grade"},"long_jump_cm":{"type":"integer","title":"Long Jump Cm"},"fifty_meter_run_ms":{"type":"integer","title":"Fifty Meter Run Ms"},"spider_ms":{"type":"integer","title":"Spider Ms"},"eight_shape_run_count":{"type":"integer","title":"Eight Shape Run Count"},"ball_throw_cm":{"type":"integer","title":"Ball Throw Cm"},"total_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Score"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["grade","long_jump_cm","fifty_meter_run_ms","spider_ms","eight_shape_run_count","ball_throw_cm","id"],"title":"MaxDataResponse"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"grade":{"type":"string","title":"Grade"}},"type":"object","required":["name","grade"],"title":"UserCreate"},"UserRead":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"grade":{"type":"string","title":"Grade"},"results":{"items":{"$ref":"#/components/schemas/UserResultRead"},"type":"array","title":"Results"}},"type":"object","required":["id","name","grade","results"],"title":"UserRead"},"UserResultCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"date":{"type":"string","format":"date","title":"Date"},"long_jump_cm":{"type":"integer","exclusiveMinimum":0.0,"title":"Long Jump Cm"},"fifty_meter_run_ms":{"type":"integer","exclusiveMinimum":0.0,"title":"Fifty Meter Run Ms"},"spider_ms":{"type":"integer","exclusiveMinimum":0.0,"title":"Spider Ms"},"eight_shape_run_count":{"type":"integer","exclusiveMinimum":0.0,"title":"Eight Shape Run Count"},"ball_throw_cm":{"type":"integer","exclusiveMinimum":0.0,"title":"Ball Throw Cm"}},"type":"object","required":["user_id","date","long_jump_cm","fifty_meter_run_ms","spider_ms","eight_shape_run_count","ball_throw_cm"],"title":"UserResultCreate"},"UserResultRead":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"date":{"type":"string","format":"date","title":"Date"},"long_jump_cm":{"type":"integer","title":"Long Jump Cm"},"fifty_meter_run_ms":{"type":"integer","title":"Fifty Meter Run Ms"},"spider_ms":{"type":"integer","title":"Spider Ms"},"eight_shape_run_count":{"type":"integer","title":"Eight Shape Run Count"},"ball_throw_cm":{"type":"integer","title":"Ball Throw Cm"}},"type":"object","required":["id","user_id","date","long_jump_cm","fifty_meter_run_ms","spider_ms","eight_shape_run_count","ball_throw_cm"],"title":"UserResultRead"},"UserUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"grade":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grade"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}